New timingsrc attribute on HTMLMediaElement to associate with TimingObject
I added a new section to the spec that proposes an extension to the definition
of "media element" to add the notion of a "current timing source" along with a
new "timingsrc" attribute to get/set the corresponding TimingObject and a few
procedures that explain how things are supposed to work.

The procedures are far from being complete but I think that's enough to
"get the idea" for now and start gathering feedback. There will be plenty
of details to look at as media elements have been defined from the ground up
to stall whenever media data is missing, which means basically all the
related procedures defined in HTML5 would need some love to introduce the
association with a TimingObject properly. There should not be anything
blocking though.