Added Sequencer section to to associate TextTrack and TimingObject
This commit proposes a new TimingTextTrack interface that inherits from
TextTrack and can be associated with a TimingObject, thus enabling the
TimingObject to act as sequencer in a very similar way to how a media
element acts as sequencer for its list of text tracks.

Introducing mechanisms similar to the ones that exists for media elements
required to pull a lot of definitions from HTML5. Some parts are incomplete
but hopefully the parts that are specified give a good idea of the technical
solution that is being proposed.

I added a warning in the Status of This Document about the fact that the spec
does not attempt to be complete at this stage.