List of use cases and requirements moved to an Appendix
I took the liberty to move the list of use cases and requirements to the
end of the spec. Now that the spec has technical content, I think that content
should be put forward right after the introduction. The list of use cases
required the reader to scroll down both in the Table of Contents or in the
spec itself to reach the right part.

It does not entail that the list of use cases is not useful! Quite the
contrary! It could perhaps be moved to a separate "Primer" or "Use cases"
document, but I think it's fine to have that list in this spec as well.

Note that I added a link to the use cases appendix in the introduction
(right before section 1.1).