Switched to ReSpec contiguous mode for WebIDL definitions
It has come to my attention that the "IDL in trees" mode that I used for the
WebIDL definitions is being deprecated from ReSpec in favor of the "contiguous
IDL" mode that is more flexible:
http://www.w3.org/respec/guide.html#webidl-support

I switched to using the contiguous mode throughout the spec as a consequence.
This has the benefit that it makes the Table of Contents simpler and avoids
a duplicate between the definition of an attribute and the procedures attached
to it.