Simplified procedure to associate a timing text track with a timing object
If we drop the notion of "list of timing text tracks", it does not make sense
to talk about "old timing" and "new timing" in the algorithm.

By the way, I note that the "time marches on" algorithm refers to the list of
text tracks of a media element, so more changes would be required to apply it
directly to timing text track objects if we don't introduce a similar notion
for timing objects. Given that the "track marches on" procedure is left
underspecified at this stage, that's not a big deal.