Updated sequencing part to separate spec from implementation
While it's a good idea to reference the JavaScript implementation from
within a draft spec, the spec is not the right place to discuss a possible
approach for things that could be up for standardisation: it *is* a
standardisation proposal.

Also, the spec should not contain example code for the JavaScript
implementation since it is not entirely aligned with the interfaces
defined. I updated the example accordingly (using HTML5.1 DataCue objects).

I re-introduced the notion of "current sequencer" for timing text tracks
instead of "current timing source". I agree that the latter would be more
consistent with the section on media element, but that creates a namespace
issue as it is tied to "media elements" in this spec, so we need another
term (or we would need to define the concept of "current timing source" more
generically).