Dropped Interval interface as agreed in #12
The TimingObject and TimingProvider interfaces now have "startPosition"
and "endPosition" properties and the prose and procedures have been updated
accordingly.

Note the "start" and "end" prefixes are not really correct since a timing
object may move in any direction along its unidimensional axis. That's
consistent with the DataCue interface in HTML5.1 though.