Separated the performance clock, the internal clock and the logical clock
The updates were made to clarify what is what.

The "performance clock" is the one exposed by Performance.now(). I also made
HR-TIME a normative dependency (instead of an informative one), since we now
require its support.

A timing object has an "internal clock" which is either the "performance clock"
or a "logical clock" that follows the skew updates reported by the timing
resource provider.

I also updated the "query()" procedure of timing object which was still
referring to the now gone "now()" method.