[Editorial] Fixes on Timing Object procedures and skew adjustments
The updates should not change the meaning. I merged the parts that talked
about gradually applying the skew adjustments (the prose was kind of
duplicated) and created a proper "calculate_skew_adjustment" in particular.

Also updated wording a bit in the TimingObject procedures (a few references
and uses of <var>) and fixed escaping of & and < characters.