Update spec to latest version of ReSpec
The spec used an outdated version of ReSpec. This update bumps the version and
refreshes cross-references accordingly.

This update also switches most HTML references to the HTML spec, with the
exception of the references to media controller that continue to target the
superseded HTML5 specification. The terminology still contains the concepts that
are defined in but not exported by HTML.